move coreclr castcache to a separate file
